requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-bootstrap4==1.0.1
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-redirect-urls==1.0
django-tables2==2.2.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.3
